Option to use a HTTP GET request instead of POST
This changed adds a new option to the Client contructor called
`http_get`, which is a boolean defaulting to False.

If this option is set to True, the Client would contruct a URL for a
HTTP GET request instead of the otherwise used POST request (determined
if a caller provides the `data` parameter or not). According to the
OAI-PMH standard both methods are allowed. But I came a long yet
another fauly implementation, where POST would not work, only GET does.
This flag is a workaround that might come in handy for others as well.